% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_itemplot.R
\name{bit_itemplot}
\alias{bit_itemplot}
\title{Plot item response curves on the bit score scale}
\usage{
bit_itemplot(model, item, bit_grid_size = 10000)
}
\arguments{
\item{model}{A fitted unidimensional mirt model object from the mirt package.}

\item{item}{A single numeric value indicating which item to plot.}

\item{bit_grid_size}{An integer specifying the internal grid size for
\code{\link{bit_scores}}. A larger value produces more accurate bit
score estimates. Default is 10000.}
}
\value{
A ggplot2 object displaying the item response curves on the bit scale.
}
\description{
This function takes a fitted unidimensional mirt model object and plots the item
response curves (category probability curves) for a given item on the bit scale.
}
\examples{
\dontrun{
library(mirt)
# Example dataset
a <- matrix(rlnorm(30, meanlog = 0, sdlog = 0.5))
d <- matrix(rnorm(30, 0, 1))
data <- simdata(a, d, 1000, itemtype = '2PL')
# Fit the model
mirt_model <- mirt(data, 1)
# Plot the response curves for item 1 on bit scale
bit_itemplot(mirt_model, item = 1)
}

}
