% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_plot_transformation.R
\name{bit_plot_transformation}
\alias{bit_plot_transformation}
\title{Plot thetas against their bit transformed counterparts}
\usage{
bit_plot_transformation(
  model,
  bit_grid_size = 10000,
  bit_scale_items = 1:extract.mirt(model, "nitems"),
  thetas_to_plot = seq(-9, 9, length.out = 501)
)
}
\arguments{
\item{model}{A fitted unidimensional mirt model object from the mirt package.}

\item{bit_grid_size}{An integer specifying the internal grid size for \code{\link{bit_scores}}. A larger value produces more accurate bit score estimates. Default is 10000.}

\item{bit_scale_items}{A numeric vector indicating which items to use for bit score computation. By default, all items are used.}

\item{thetas_to_plot}{The range of theta values to plot the test information curve. The bit scores and the information is computed for these values. Default is from -9 to 9 with 501 points.}
}
\value{
A ggplot2 object displaying the item response curves on the bit scale.
}
\description{
Plot thetas against their bit transformed counterparts
}
\examples{
\dontrun{
library(mirt)
# Example dataset
a <- matrix(rlnorm(30, meanlog = 0, sdlog = 0.5))
d <- matrix(rnorm(30, 0, 1))
data <- simdata(a, d, 1000, itemtype = '2PL')
# Fit the model
mirt_model <- mirt(data, 1)
# Plot the response curves for item 1 on bit scale
bit_plot_transformation(mirt_model)
}

}
