% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_score_information.R
\name{bit_score_information}
\alias{bit_score_information}
\title{Approximate the bit score Fisher information for a given mirt model at the bit scale locations corresponding to the supplied theta scores.}
\usage{
bit_score_information(
  model,
  theta,
  item = NULL,
  bit_scale_items = 1:extract.mirt(model, "nitems")
)
}
\arguments{
\item{model}{An fitted mirt model object from the mirt package}

\item{theta}{A vector or matrix with theta scores. Typically the matrix returned from the mirt::fscores method.}

\item{item}{If provided, the information provided by the specified item is computed as opposed to all items.}

\item{bit_scale_items}{A numeric vector indicating which items to use for bit score computation. By default, all items are used.}
}
\value{
An information vector for each element in thetas.
}
\description{
Approximate the bit score Fisher information for a given mirt model at the bit scale locations corresponding to the supplied theta scores.
}
\examples{
\dontrun{
library(mirt)
# As an example, we simulate a dataset with 30 items and 1000 respondents
# discrimination parameters from a log-normal distribution
a <- matrix(rlnorm(30, meanlog = 0, sdlog = 0.5))
# difficulty parameters from a standard normal distribution
d <- matrix(rnorm(30, 0, 1))
data <- simdata(a, d, 1000, itemtype = '2PL')
# Fit the model and compute theta scores
mirt_model <- mirt(data, 1)
theta <- fscores(mirt_model, full.scores.SE = FALSE)
# Compute the bit score test information
test_info <- bit_score_information(mirt_model, theta)
# Compute the bit score item information for item 3
item_info <- bit_score_information(mirt_model, theta, item = 3)
}

}
